Modify Nutrient/Water System to use its map to store changes by other systems
Really just using the map as a queue but it provides mechanism for other
systems(survival) to make changes while not allowing them to directly modify vectors.

Might want to rethink design by just having the map store a copy of every
water/nutrient value. Current "queue" design would likely encounter seg faults when map
is cleared if ran concurrently with survivalSystem.  With a map copy, survivalSystem
could write to the map while water/nutrient systems write to vectors. The race conditions
in this case of reading from map to update vectors would be negligible in this case.
This would require initialization of the map value after a component is emplaced.